feat: adding logo; redefining project strcuture and landing page; reducing and simplifying some components
