+++ next.config.mjs
@@ -1,4 +1,14 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  async redirects() {
+    return [
+      {
+        source: "/",
+        destination: "/advocates",
+        permanent: true,
+      },
+    ];
+  },
+};
 
 export default nextConfig;
