+++ package.json
@@ -26,6 +26,7 @@
     "eslint": "^8.57.0",
     "eslint-config-next": "14.2.5",
     "postcss": "^8.4.40",
+    "prettier": "^3.4.2",
     "tailwindcss": "^3.4.7",
     "typescript": "^5.5.4"
   }
