+++ src/app/layout.tsx
@@ -1,12 +1,18 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
+import { Poppins } from "next/font/google";
+import { Header } from "@/components/header";
+
 import "./globals.css";
 
-const inter = Inter({ subsets: ["latin"] });
+const poppins = Poppins({
+  subsets: ["latin"],
+  display: "swap",
+  weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"],
+});
 
 export const metadata: Metadata = {
   title: "Solace Candidate Assignment",
-  description: "Show us what you got",
+  description: "2 hours goes fast",
 };
 
 export default function RootLayout({
@@ -16,7 +22,10 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={poppins.className}>
+        <Header />
+        <main className="flex flex-col items-center">{children}</main>
+      </body>
     </html>
   );
 }
